pyproject.toml
[project]
name = "albench-s2f"
version = "0.1.0"
description = "Active learning benchmark for genomic sequence-to-function prediction"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "torch>=2.0.0",
  "numpy>=1.24.0",
  "scipy>=1.10.0",
  "pandas>=2.0.0",
  "pyyaml>=6.0",
  "tqdm>=4.65.0",
  "python-dotenv>=1.0.0",
  "matplotlib>=3.7.0",
  "seaborn>=0.12.0",
  "scikit-learn>=1.3.0",
  "biopython>=1.86",
  "hydra-core>=1.3.2",
  "omegaconf>=2.3.0",
  "wandb>=0.19.0",
  "imageio>=2.28.0",
  "jax>=0.4.0",
  "jaxlib>=0.4.0",
  "optax>=0.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.9.0",
  "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["albench", "data", "models", "evaluation", "experiments", "scripts"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501", "E712", "F401", "F541"]

[tool.ruff.lint.per-file-ignores]
# jaxtyping annotations like Float[Array, "B S 4"] trigger false-positive F722/F821
# F841: unused variables seq_len / logo are intentional (logomaker side-effects)
"custom_model.py" = ["F722", "F821", "F841"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyright]
venvPath = "."
venv = ".venv"
typeCheckingMode = "basic"
reportMissingImports = false

[tool.pytest.ini_options]
testpaths = ["tests"]
